requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
easyocr==1.4.2
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
idna==3.3
imageio==2.19.0
importlib-metadata==4.11.3
Jinja2
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
networkx
numpy
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.3.0.38
packaging==21.3
pandas
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy
seaborn==0.11.2
six==1.16.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile
torch
torchaudio
torchvision
tqdm
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
fastapi
uvicorn
gunicorn==19.9.0
uvloop
httptools

#FastAPI
#opencv-python==4.3.0.38